# Real Estate Web RD/api_contracts.openapi.yaml
openapi: 3.1.0
info:
  title: Real-Estate AI Web App API
  version: 0.1.0
servers:
  - url: https://clientdomain.com/api/v1
security: []
paths:
...
  /admin/content:
    get:
      summary: List admin content items
      responses:
        '200':
          description: Content items
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ContentItem' }
    post:
      summary: Create draft admin content item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentCreateRequest' }
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                type: object
                required: [item]
                properties:
                  item: { $ref: '#/components/schemas/ContentItem' }
  /admin/content/{id}:
    get:
      summary: Get admin content item
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Content item }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Patch admin content item
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentPatchRequest' }
      responses:
        '200':
          description: Content item patched
          content:
            application/json:
              schema:
                type: object
                required: [item]
                properties:
                  item: { $ref: '#/components/schemas/ContentItem' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/content/{id}/publish:
    post:
      summary: Publish content and index Phase 1A RAG chunks using stub/hash-v1
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Published and indexed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/rag/documents/{id}/chunks:
    get:
      summary: List RAG chunks for document
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: RAG chunks without raw vectors
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RagChunk' }
...
  /admin/rag/search-test:
    post:
      summary: Test pgvector RAG retrieval with stub/hash-v1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                topK: { type: integer }
      responses:
        '200':
          description: Retrieval results
          content:
            application/json:
              schema:
                type: object
                required: [query, sources, retrieval]
                properties:
                  query: { type: string }
                  sources:
                    type: array
                    items: { $ref: '#/components/schemas/RagSource' }
                  retrieval: { $ref: '#/components/schemas/RetrievalTrace' }
...
  /admin/chat-lab/sessions/{id}/messages:
    get:
      summary: List chat lab messages
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Chat lab transcript
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ChatMessage' }
  /admin/chat-lab/test-message:
    post:
      summary: Send admin lab test message and get grounded answer or fallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                sessionId: { type: string, format: uuid }
                message: { type: string, minLength: 1 }
                topK: { type: integer }
      responses:
        '200':
          description: Grounded answer or fallback
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatAnswerResponse' }
...
  /webhooks/meta/whatsapp:
    get:
      summary: Verify WhatsApp webhook subscription
      parameters:
        - in: query
          name: hub.mode
          schema: { type: string }
        - in: query
          name: hub.challenge
          schema: { type: string }
        - in: query
          name: hub.verify_token
          schema: { type: string }
      responses:
        '200': { description: Webhook challenge response }
    post:
      summary: Receive WhatsApp Business Platform webhook events
      responses:
        '202':
          description: Event accepted and processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetaWebhookProcessResponse' }
  /webhooks/meta/instagram:
    get:
      summary: Verify Instagram Messaging webhook subscription
      parameters:
        - in: query
          name: hub.mode
          schema: { type: string }
        - in: query
          name: hub.challenge
          schema: { type: string }
        - in: query
          name: hub.verify_token
          schema: { type: string }
      responses:
        '200': { description: Webhook challenge response }
    post:
      summary: Receive Instagram Messaging webhook events
      responses:
        '202':
          description: Event accepted and processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetaWebhookProcessResponse' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    MetaWebhookProcessResponse:
      type: object
      required: [accepted, channel, processed, replied, skipped, failures]
      properties:
        accepted: { type: boolean }
        channel: { type: string, enum: [whatsapp, instagram_dm] }
        processed: { type: integer }
        replied: { type: integer }
        skipped: { type: integer }
        failures:
          type: array
          items:
            type: object
            required: [messageId, detail]
            properties:
              messageId: { type: string }
              detail: { type: string }
...
    ContentCreateRequest:
      type: object
      required: [title, body]
      properties:
        contentType: { type: string, enum: [project, property, faq, area, policy, general] }
        title: { type: string }
        slug: { type: string }
        body: { type: string }
        metadata: { type: object }
    ContentPatchRequest:
      type: object
      properties:
        contentType: { type: string, enum: [project, property, faq, area, policy, general] }
        title: { type: string }
        slug: { type: string }
        body: { type: string }
        metadata: { type: object }
    ContentItem:
      type: object
      required: [id, contentType, title, slug, body, status]
      properties:
        id: { type: string, format: uuid }
        contentType: { type: string }
        title: { type: string }
        slug: { type: string }
        body: { type: string }
        status: { type: string, enum: [draft, published] }
        publishedVersionId: { type: [string, 'null'], format: uuid }
    PublishResponse:
      type: object
      required: [item, versionId, documentId, chunkCount, embeddingModel]
      properties:
        item: { $ref: '#/components/schemas/ContentItem' }
        versionId: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        chunkCount: { type: integer }
        embeddingModel: { type: string, enum: [stub/hash-v1, BAAI/bge-m3] }
...
    RagChunk:
      type: object
      required: [id, documentId, content, embeddingModel, embeddingDimension]
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        content: { type: string }
        embeddingModel: { type: string, enum: [stub/hash-v1, BAAI/bge-m3] }
        embeddingDimension: { type: integer, enum: [1024] }
    RagSource:
      type: object
      required: [chunkId, title, excerpt, score]
      properties:
        chunkId: { type: string, format: uuid }
        title: { type: string }
        excerpt: { type: string }
        score: { type: number }
    RetrievalTrace:
      type: object
      required: [topK, embeddingModel]
      properties:
        topK: { type: integer }
        embeddingModel: { type: string, enum: [stub/hash-v1, BAAI/bge-m3] }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        requestId: { type: string }
...
    ChatMessage:
      type: object
      required: [id, conversationId, direction, messageType, content, sourceIds, actionTrace, createdAt]
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        direction: { type: string, enum: [inbound, outbound, system] }
        messageType: { type: string }
        content: { type: string }
        sourceIds:
          type: array
          items: { type: string }
        actionTrace: { type: object }
        createdAt: { type: string, format: date-time }
    ChatAnswerResponse:
      type: object
      required: [answer, sources, fallback, retrieval, leadScore, handoff]
      properties:
        sessionId: { type: string, format: uuid }
        answer: { type: string }
        fallback: { type: boolean }
        sources:
          type: array
          items: { $ref: '#/components/schemas/RagSource' }
        retrieval: { $ref: '#/components/schemas/RetrievalTrace' }
        leadScore:
          type: object
          required: [score, band, reasonCodes]
          properties:
            score: { type: integer, enum: [0] }
            band: { type: string, enum: [cold] }
            reasonCodes:
              type: array
              items: { type: string }
        handoff: { type: 'null' }
...